adjust

+++ main.go
@@ -118,9 +118,9 @@ func replaceImage(originImage string, fileName string) string {
 
 		originImage = path + originPath + string(os.PathSeparator) + originImage
 
-		remote, _ = client.UploadFile(originImage, remoteName)
+		_, err := client.UploadFile(originImage, remoteName)
 
-		if os.Getenv("DELETE_ORIGIN_URL") != "" {
+		if os.Getenv("DELETE_ORIGIN_URL") != "" && err != nil {
 			log.Printf("删除原有图片文件 <%s>", originImage)
 			os.Remove(originImage)
 		}
